LP1849523 Angular catalog split record batches
The result page of the Angular catalog now fetches the record info in 2
batches.  The first batch is relatively small (currently 5 records) and
allows the results page to start rendering results earlier in the
display process.  The second batch of records, which includes all the
rest of the results, then drops in as they arrive.

Note the first batch of records will be the first 5 results from the
search so the second batch of results can simply be appended and avoid
page shuffling.

Includes thinko fix for batched holds retrieval.

Signed-off-by: Bill Erickson <berickxx@gmail.com>